+++ hellion-cli/src/main/java/module-info.java
@@ -1,0 +1,8 @@
+module org.monarchinitiative.hellion.cli {
+    requires org.monarchinitiative.hellion.core;
+    requires org.monarchinitiative.biodownload;
+    requires info.picocli;
+    requires org.slf4j;
+
+    opens org.monarchinitiative.hellion.cli.cmd to info.picocli;
+}
