+++ hellion-core/src/main/java/org/monarchinitiative/hellion/core/FuzzStrategy.java
@@ -1,0 +1,10 @@
+package org.monarchinitiative.hellion.core;
+
+public enum FuzzStrategy {
+
+    ADD_N_RANDOM_HPO_TERMS,
+    REPLACE_HPO_WITH_PARENT,
+    REPLACE_HPO_WITH_GRANDPARENT,
+    DROP_ONE_CAUSATIVE_VARIANT
+
+}
