+++ hellion-core/src/main/java/org/monarchinitiative/hellion/core/PhenopacketFuzzer.java
@@ -1,0 +1,9 @@
+package org.monarchinitiative.hellion.core;
+
+import org.phenopackets.schema.v2.Phenopacket;
+
+public interface PhenopacketFuzzer {
+
+    Phenopacket fuzz(Phenopacket pp);
+
+}
