+++ hellion-core/src/test/resources/org/monarchinitiative/hellion/core/README.md
