+++ hellion-core/src/test/resources/org/monarchinitiative/hellion/core/hpo_toy.json
