--- phenopacket-fuzzer-core/src/main/java/module-info.java
@@ -1,10 +1,0 @@
-module org.monarchinitiative.phenopacket_fuzzer.core {
-    requires transitive org.phenopackets.schema;
-    requires org.phenopackets.phenopackettools.builder;
-    requires org.monarchinitiative.phenol.core;
-    requires org.monarchinitiative.phenol.annotations;
-    requires org.slf4j;
-
-    exports org.monarchinitiative.phenopacket_fuzzer.core;
-
-}
