--- phenopacket-fuzzer-core/src/main/java/org/monarchinitiative/phenopacket_fuzzer/core/PhenopacketFuzzer.java
@@ -1,9 +1,0 @@
-package org.monarchinitiative.phenopacket_fuzzer.core;
-
-import org.phenopackets.schema.v2.Phenopacket;
-
-public interface PhenopacketFuzzer {
-
-    Phenopacket fuzz(Phenopacket pp);
-
-}
